i915/gem_exec_whisper: Trim the queue length
Everything was 1024x1024 originally simply for that being a nice round
number. However, when running with multiple processes each submitting
a chain of 1024 for 1024 passes become quite taxing on the system, such
that even a single pass takes minutes. In respect of CI's motto of run
fast, run often, trim the workload substantially, leaving enough of a
chain to tax the scheduler.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>