runner: Only claim the test was killed if it was killed
If we don't have --abort=taint active and there is a kernel taint,
test exiting normally caused the runner to inject a "this test was
killed" message to the test's output. Make sure we only inject that if
we really did kill the test, and journal the test exit correctly as
well.

Same goes for the message for exceeding disk usage limits.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Arkadiusz Hiler <arek@hiler.eu>
Cc: Lukasz Fiedorowicz <lukasz.fiedorowicz@intel.com>
Reviewed-by: Lukasz Fiedorowicz <lukasz.fiedorowicz@intel.com>