lib/i915/perf: rework perf counters equations
Working with an offset means we can have the PERFCNT snapshots stored
at an offset in the existing accumulator rather than having a specific
accumulator just for them.

This isn't really useful for IGT because it only looks at the OA
buffer where those registers aren't stored, but it is useful for perf
queries in Mesa.

This bumps the version of the i915_perf library to reflect the API
change (see perf.h).

v2: Also bump so version

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Acked-by: Petri Latvala <petri.latvala@intel.com>