tests/kms_big_joiner: Add test to validate big joiner
Added negative test to verify the different scenarios for big joiner.
In the test, modeset is done on Pipe A for a big joiner mode and
a second modeset is attempted on Pipe B which is expected to fail.
Same functionality is validated for other pipes as well.

Secondly, the reverse is tested where a modeset is done on Pipe B
and then a second big joiner modeset is attempted on Pipe A which is
expected to fail. Same functionality is validated for other pipes as well.

v2: -Rename HDISPLAY_5K to MAX_HDISPLAY_PER_PIPE. (Manasi)
    -Do not assume mode[0] will be the highest mode.
     Loop through the mode list to find the 8k mode. (Manasi)
    -Add a subtest for verifying basic 8k modeset on all pipes. (Manasi)

v3: -Add break after big joiner output is found in basic subtest. (Manasi)
    -Fix the clean up part in the invalid-modeset subtest.

v4: -Add comments in the clean up part of invalid subtest as to why only
     particular fb needs to be set to NULL. (Manasi)

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Manasi Navare <manasi.d.navare@intel.com>