tests/kms_async_flips: Modify assert to fix failure seen on few platforms
Failure is seen on CI on few platforms as one of the flips in between is
taking longer time. To fix this, added a check to assert for individual
flips only for the alternate-sync-async-flip subtest. In the
async-flip-with-page-flip-events subtest we anyway check the overall FPS
at the end, so we do not need a check for individual flips and this should
take care of some random flip taking more time in between.

Also, reduced the MIN_FLIPS_PER_FRAME from 8 to 5 based on the failure
seen in CI.

Bug: https://gitlab.freedesktop.org/drm/intel/-/issues/2521

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Kunal Joshi <kunal1.joshi@intel.com>