tests/kms_dp_aux_dev: Handle AUX failures on disconnected MST connectors
The DPCD of an MST connector is read out with a REMOTE_DPCD_READ MST
request. If the given connector is disconnected this read will result in
an MST NAK reply and this will be reported as an EIO error to the
initiator of the AUX read.

Handle this in the test that attempts to read the DPCD of any exposed
connector, whether they are connected or not.

Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>