lib/igt_map: Adopt Mesa hash table
The _search function has been changed to return a pointer to
the stored data instead of the entry struct. The _search_entry function,
which acts as the original search has been added. Additionally _remove function
has an optional delete_function param, to make it more usable.

For more information, see:
http://cgit.freedesktop.org/~anholt/hash_table/tree/README

Signed-off-by: Dominik Grzegorzek <dominik.grzegorzek@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Jason Ekstrand <jason@jlekstrand.net>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Acked-by: Petri Latvala <petri.latvala@intel.com>