tests/i915/perf: verify reason field in OA reports is never 0
We're about to remove the filtering in i915 on 0 reason fields because
we assume this was a possibility, but it turned out to be a corruption
in the tail pointer register that made us read cleared data.

This test is here to verify that our assumption hold that the HW never
produces such reports.

v2: Fix len checking (Umesh)
    Count report lost events (Umesh)
    Check report sanity (Umesh)
    Limit test to 3 times the OA buffer (Umesh)

v3: Bump OA sampling frequency to 20us, seeing some lost buffer
    failures on KBL (Lionel)

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com> (v2)