i915/api_intel_bb: Only assert objects are unmoved for full-ppgtt
If we let an object idle in a shared GTT, it may be evicted by the
kernel in favour of another client. Thus, we have to be very careful
when asserting that two different executions of the same object will
be at the same address. If there's an idle point between the two
asserts, it will only be guaranteed to hold for full-ppgtt.

Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/2810
References: c20db20a7cd7 ("i915/api_intel_bb: Only assert objects are unmoved for full-ppgtt")
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Acked-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>