kms_atomic_transition:Convert nx-modeset-transitions subtest to dynamic
In kms_atomic_transition subtest,it will test all the
pipes defined in IGT, i.e. IGT_MAX_PIPES whether
the output is available or not.Later it has to be
analysed and discarded as valid skip.To save this
time, updated the test to first check the available
outputs and then execute the test.

V7: -Modified commit message. (Karthik)
    -Replaced for_each_single_pipe_with_single__output with
     for_each_connected_output to execute the test only on
     connected display. (Karthik)

v8: -Modified the description subject line. (Petri)
    -Modified subtests names to remove redundancy. (Petri)
    -Added extra line before for loop. (Karthik)

v9: -Made separate subtests for non-blocking and
     fencing parameters. (Petri)

Signed-off-by: Nidhi Gupta <nidhi1.gupta@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>