i915/gem_tiled_w[bc]: Require some fences
In these tests we set the object via the GTT with an aperture detiling
fence, and then read them back with WC/WB cpu maps, verifying that the
read back is a linear view of the tiled buffer. These tests require both
GTT and a fence through which to detile, and are simply irrelevant if
there is no aperture detiling on the platform.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>