meson: Turn off -Werror
No, we don't want all warnings ever to be errors. While being rid of
compile warnings (especially the more obscure ones from future
compilers) is great, having it on always is not acceptable. It's not
even on "by default", it was non-overrideably set on.

Not to mention adding -Werror to compiler flags is the incorrect way
of turning it on. Use the meson command line flag --werror instead on
whatever local checkout or company-internal fork you're using, where
you have control over which compilers are used. Turning it on
unconditionally in an upstream repository for all current and future
compilers is never ok.

This partially reverts commit e02612921a4e95aef3a368e7468f4337c9dcee7d.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Ashutosh Dixit <ashutosh.dixit@intel.com>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Arkadiusz Hiler <arek@hiler.eu>
Reviewed-by: Chris Wilson <chris@chris-wilson.co.uk>