tests/kms_atomic_transition: correctly restrict execution to 2 pipes
In multi display scenario, the subtest will get executed on the
first pipe and exit,and does not execute on the second pipe as required.
Added a change to take into account the number of the outputs connected,
so that the subtest is executed on all the outputs, on the first 2 pipes.

Signed-off-by: Nidhi Gupta <nidhi1.gupta@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>