tests/kms-hdr: Fix HDMI_STATIC_METADATA_TYPE1.
According to the CTA 861.G specification, table 44
in section 6.9 and also specifically section 6.9.1,
the value of the constant must be 0, not 1, so fix
this.

An identical fix has been submitted for inclusion into
the kernel headers (this is not exposed uapi). See:

https://lists.freedesktop.org/archives/dri-devel/2021-February/297233.html

Fixing this constant for Kodi also fixed some HDR TVs,
so the wrong constant has real world consequences on
some display hardware. See:

https://github.com/xbmc/xbmc/pull/19129

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>
Cc: Simon Ser <contact@emersion.fr>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>