amdgpu/info: convert asserts to human-readable explanations
Having a test fail with "r != 0" is not helpful, and forces developers
to open the source code of IGT to figure out what this actually means.

This commit changes these asserts so as the function call that failed
is in the stack trace, and adds a human-readable explanation of what
failed.

Suggested-by: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Martin Peres <martin.peres@mupuf.org>
Reviewed-by: Bas Nieuwenhuizen <bas@basnieuwenhuizen.nl>