tests/kms_vrr: Extend to cover full VRR range
As we are restricting the vrr max to current mode vrefresh,
we are not covering the full vrr range with default mode.

So, instead of running tests on default mode, loop through the
all connector modes and find the mode with max refresh rate to
exercise full vrr range.

V2:
* Add a check if vrr vmax >= connector mode vrefresh (Manasi)

Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Cc: Latvala Petri <petri.latvala@intel.com>
Suggested-by: Manasi Navare <manasi.d.navare@intel.com>
Signed-off-by: Bhanuprakash Modem <bhanuprakash.modem@intel.com>
Reviewed-by: Manasi Navare <manasi.d.navare@intel.com>