tests/vc4_dmabuf_poll: Fix checking for POLLOUT
This test has two subtests, one which polls for POLLIN on the dmabuf fd,
and the other which looks like it's supposed to poll for POLLOUT. They
are both implemented with a helper function that takes the polling flag
as an argument. It looks like the helper actually ignores the argument
and both subtests end up checking for POLLIN. This just fixes it to use
the argument.

Fixes: da42b17e45c1 ("igt/vc4_dmabuf_poll: Add a test for polling to wait for dmabuf")
Signed-off-by: Neil Roberts <nroberts@igalia.com>
Reviewed-by: Eric Anholt <eric@anholt.net>