tests/kms_atomic_transition: Reset the pipe_count variable
The pipe_count variable is not being reset after every subtest.
This is not causing any issue when running the subtests individually,
but when running the full test only the first subtest runs
and all the other subtests get skipped.

To fix this, reset the pipe_count variable after every subtest.

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Kunal Joshi <kunal1.joshi@intel.com>