tests/kms_atomic_transition: Enable Primary Plane on Enabling CRTC
It is common to find drivers that require the primary plane enabled
whenever the CRTC is (almost all drm_tiny drivers and amdgpu need this).
Nevertheless, kms_atomic_transition does not consider this case and, for
this reason, fails in hardware that has the aforementioned restriction.
Since this is not part of the test, this commit ensures that the test
setup all plane params before committing anything. This commit was
tested on: Delbin, Volteer, Raven, and VKMS.

Changes since V1:
- Update commit message
- Rebase

Cc: Harry Wentland <harry.wentland@amd.com>
Cc: Nicholas Choi <Nicholas.Choi@amd.com>
Cc: Sean Paul <seanpaul@chromium.org>
Cc: Mark Yacoub <markyacoub@chromium.org>
Cc: Melissa Wen <melissa.srw@gmail.com>
Cc: Arkadiusz Hiler <arek@hiler.eu>
Cc: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Mark Yacoub <markyacoub@chromium.org>
Signed-off-by: Rodrigo Siqueira <Rodrigo.Siqueira@amd.com>
Reviewed-By: Nicholas Choi <Nicholas.Choi@amd.com>