tests/i915/gem_watchdog: Exercise long rendering chains
Test to demonstrate a problem with the proposed default fence expiry
semantics where long rendering chain get silently broken.

If we had fence error propagation (no clear agreement whether to do it or
not) maybe userspace would see if, assuming fence status is looked at, but
overall potential rendering corruption is the story in any case.

Note that this is not a single long batch but just a long queue of work
which could be viewed as heavy system load as well (like virtualisation
or other types of resource sharing).

v2:
 * Fix commit message.
 * Blacklist the test which is expected to fail for now.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>