Revert "Add eDP support in Link Layer Compliance Tool"
The eDP change didn’t take into account choosing the right CRTC when compliance is requested on external DP.
This causes Hotplug Uevent sent from the kernel (intel_dp_short_pulse) to always trigger LT on eDP.
Without reverting this change the only way to run compliance on external display is to disconnect the eDP cable which is pain for ODMs.

So reverting this change until a full solution that handles multiple connected displays in place.

This reverts commit 391929b7402fd0417dcb9026c202241413653942.

Signed-off-by: Khaled Almahallawy <khaled.almahallawy@intel.com>
Acked-by: Petri Latvala <petri.latvala@intel.com>