lib/intel_bb: Remove intel_bb_assign_vm and tests (v2)
It's not used by anything other than the tests for that functionality
and it relies on setting the VM via SET_CONTEXT_PARAM which is
deprecated.  Delete it for now.  We can add it back in later if it's
actually useful and do it properly then.

v2 (Zbigniew Kempczyński):
 - Also remove intel_bb_detach_intel_buf()

Signed-off-by: Jason Ekstrand <jason@jlekstrand.net>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Acked-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>