lib/igt_fb: Remove redundant VC4 SAND plane size alignment
Special casing the tile_height > plane_height by rounding up the
plane_height to tile_height is redundant, since in that case the
following return statement will do exactly the same. Remove the special
casing.

The only exception would be if plane_height was 0, but the commit adding the
special case doesn't mention this as a possiblity and I presume that as an
framebuffer configuration.

Cc: Paul Kocialkowski <paul.kocialkowski@bootlin.com>
Cc: Lyude Paul <lyude@redhat.com>
Cc: Maxime Ripard <maxime.ripard@bootlin.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: Juha-Pekka Heikkilä <juha-pekka.heikkila@intel.com>