tests/kms_dp_dsc: Bug fix for DP on Pipe A
DSC is not supported on Pipe A for external DP on Gen11 platforms.
But as this check is within update_display(), we still call test_cleanup().
This bug is seen here as the pipe wasn't set properly for this output.

To fix this, moving the check outside update_display() so that we do not
call test_cleanup if we didn't even run the test.

Also, moved the fb creation and removal outside the loop for_each_pipe(),
since it seems redundant within the loop.

Fixes: https://gitlab.freedesktop.org/drm/intel/-/issues/3466

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Manasi Navare <manasi.d.navare@intel.com>