lib/i915: Add ADL-P stride restrictions for non-linear buffers
ADL-P tiled framebuffer strides must be power-of-two aligned and has a
minimum of 8 tiles. For non-CCS framebuffers the driver supports FBs not
meeting this requirement by remapping the framebuffer and padding the
stride as required, but for CCS FBs userspace must ensure the alignment.
Adding remap support for CCS FBs to the driver is to be done as a
follow-up.

Cc: Juha-Pekka Heikkilä <juha-pekka.heikkila@intel.com>
Cc: Anusha Srivatsa <anusha.srivatsa@intel.com>
Cc: Clinton Taylor <Clinton.A.Taylor@intel.com>
Cc: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Cc: Imre Deak <imre.deak@intel.com>
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Imre Deak <imre.deak@intel.com>