tests/gem_exec_schedule: Use store_dword_plug again (v2)
In

commit 2884f91dd6d7682ea738ef6f0943cc591643dda2
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Fri Jul 20 12:29:30 2018 +0100

    igt/gem_exec_schedule: Trim deep runtime

this was open-coded, I guess to avoid having to allocate a ton of
batchbuffers. Unfortunately this relies on being able to rewrite
batchbuffer relocations while the batch is in-flight (otherwise
everything stalls and our setup is for nothing), and we're removing
gpu relocations from upstream.

This problem was realized for the testcase in general in

commit f1e62e330a6e2de7b3cbf7cf02d71ae00cc6adcc
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Tue May 26 15:22:38 2020 +0100

    i915/gem_exec_schedule: Reduce relocation risk for store-dword

but the fix in there is only stochastic, there's still a chance of
failure and hence unsightly noise in CI. The proper fix is to use
softpin for this testcase unconditionally (not just when relocations
aren't available), so that we clearly disentangle the concerns here
and really only test the scheduler. And not some legacy features that
are on the way out to their deserved retirement like relocations.

Zbyscek and Ashotush cc'ed so they're aware that this testcase must be
switched over to softping uncondtionally, for correctness reasons.

But that's a pile more work, so as an interim solution go back to
__store_dword helper.

v2 (Jason Ekstrand):
 - Rebase on intel_ctx_t changes

Cc: Jason Ekstrand <jason@jlekstrand.net>
Cc: Dixit Ashutosh <ashutosh.dixit@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Cc: Dave Airlie <airlied@redhat.com>
Cc: Lakshminarayana Vudum <lakshminarayana.vudum@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>
Reviewed-by: Jason Ekstrand <jason@jlekstrand.net>