lib/igt_fb: Add support for testing of 16 bpc fixed point formats.
This is used to support testing the 16 bpc formats, e.g., via:

kms_plane --run-subtest pixel-format-pipe-A-planes

So far this was successfully tested on AMD RavenRidge with DCN-1
display hw. Ville also tested it against some of his patches for
16 bpc support on i915.

The new conversion routines are slightly adapted copies of the
convert_float_to_fp16() and convert_fp16_to_float() functions,
with the conversion math modified for float <-> uint16 instead.

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: Alex Deucher <alexander.deucher@amd.com>